\name{Ones}
\alias{Ones}

\title{
Matrix of ones
}
\description{
Square matrix of ones
}
\usage{
Ones(n)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{n}{
Order of the matrix.
}
}
\details{
Square matrix of ones.
}
\value{
A matrix of ones of order n.
}

\author{
Jose Luis Vicente-Villardon
}

\examples{
Ones(6)
}
