\name{spiders}
\alias{spiders}
\docType{data}
\title{
Hunting Spiders Data
}
\description{
Hunting spiders data transformed into Presence/Absence.
}
\usage{data("spiders")}
\format{
  A data frame with 28 observations of presence/absence on the following 12 hunting spider species.
  \describe{
    \item{\code{Alopacce}}{a factor with levels \code{Absent} \code{Present} of the species Alopecosa accentuata.}
    \item{\code{Alopcune}}{a factor with levels \code{Absent} \code{Present} of the species Alopecosa cuneata.}
    \item{\code{Alopfabr}}{a factor with levels \code{Absent} \code{Present} of the species Alopecosa fabrilis.}
    \item{\code{Arctlute}}{a factor with levels \code{Absent} \code{Present} of the species Arctosa lutetiana.}
    \item{\code{Arctperi}}{a factor with levels \code{Absent} \code{Present} of the species Arctosa perita.}
    \item{\code{Auloalbi}}{a factor with levels \code{Absent} \code{Present} of the species Aulonia albimana.}
    \item{\code{Pardlugu}}{a factor with levels \code{Absent} \code{Present} of the species Pardosa lugubris.}
    \item{\code{Pardmont}}{a factor with levels \code{Absent} \code{Present} of the species Pardosa monticola.}
    \item{\code{Pardnigr}}{a factor with levels \code{Absent} \code{Present} of the species Pardosa nigriceps.}
    \item{\code{Pardpull}}{a factor with levels \code{Absent} \code{Present} of the species Pardosa pullata.}
    \item{\code{Trocterr}}{a factor with levels \code{Absent} \code{Present} of the species Trochosa terricola.}
    \item{\code{Zoraspin}}{a factor with levels \code{Absent} \code{Present} of the species Zora spinimana.}
  }
}

\source{
van der Aart, P. J. M., and Smeenk-Enserink, N. (1975) Correlations between distributions of hunting spiders (Lycos- idae, Ctenidae) and environmental characteristics in a dune area. Netherlands Journal of Zoology 25, 1-45.
}

\examples{
data(spiders)
}
\keyword{datasets}
